fix: bio-nlp to bc5dr toggle functionality
